+++ setup.sh
@@ -24,38 +24,66 @@ if [ $USER = "root" ]; then
   exit 1
 fi
 
+if whiptail --yesno "Would you like to enable ps3netsrv for PS3 support? (SMB is enabled either way for PS2 support etc.)" 8 55; then
+  PS3NETSRV=true
+else
+  PS3NETSRV=false
+fi
+
+if whiptail --yesno "Would you like to enable XLink Kai?" 8 55; then
+  XLINKKAI=true
+else
+  XLINKKAI=false
+fi
+
+if whiptail --yesno "Would you like to enable wifi access point for a direct wifi connection?" 8 55; then
+  WIFIACCESSPOINT=true
+else
+  WIFIACCESSPOINT=false
+fi
+
 # Update packages
 sudo apt-get -y update
 sudo apt-get -y upgrade
 
 # Ensure basic tools are present
-sudo apt-get -y install screen wget git curl coreutils
+sudo apt-get -y install screen wget git curl coreutils iptables hostapd
 
 # Install and configure Samba
 sudo apt-get install -y samba samba-common-bin
-wget https://raw.githubusercontent.com/toolboc/psx-pi-smbshare/master/samba-init.sh -O /home/${USER}/samba-init.sh
+wget https://raw.githubusercontent.com/georgewoodall82/psx-pi-smbshare-updated/master/samba-init.sh -O /home/${USER}/samba-init.sh
+sed -i "s/userplaceholder/${USER}/g" /home/${USER}/samba-init.sh
 chmod 755 /home/${USER}/samba-init.sh
 sudo cp /home/${USER}/samba-init.sh /usr/local/bin
 sudo mkdir -m 1777 /share
 
-# Install ps3netsrv
-sudo rm /usr/local/bin/ps3netsrv++
-sudo apt-get install -y git gcc
-git clone https://github.com/dirkvdb/ps3netsrv--.git
-cd ps3netsrv--
-git submodule update --init
-make CXX=g++
-sudo cp ps3netsrv++ /usr/local/bin
+# Install ps3netsrv if PS3NETSRV is true
+if [ "$PS3NETSRV" = true ]; then
+  sudo rm /usr/local/bin/ps3netsrv++
+  sudo apt-get install -y git gcc
+  git clone https://github.com/dirkvdb/ps3netsrv--.git
+  cd ps3netsrv--
+  git submodule update --init
+  make CXX=g++
+  sudo cp ps3netsrv++ /usr/local/bin
+fi
 
 # Install wifi-to-eth route settings
 sudo apt-get install -y dnsmasq
-wget https://raw.githubusercontent.com/toolboc/psx-pi-smbshare/master/wifi-to-eth-route.sh -O /home/${USER}/wifi-to-eth-route.sh
+wget https://raw.githubusercontent.com/georgewoodall82/psx-pi-smbshare-updated/master/wifi-to-eth-route.sh -O /home/${USER}/wifi-to-eth-route.sh
 chmod 755 /home/${USER}/wifi-to-eth-route.sh
 
-# Install setup-wifi-access-point settings
-sudo apt-get install -y hostapd bridge-utils
-wget https://raw.githubusercontent.com/toolboc/psx-pi-smbshare/master/setup-wifi-access-point.sh -O /home/${USER}/setup-wifi-access-point.sh
+if [ "$WIFIACCESSPOINT" = true ]; then
+  # Install setup-wifi-access-point settings
+  sudo apt-get install -y hostapd bridge-utils
+  wget https://raw.githubusercontent.com/georgewoodall82/psx-pi-smbshare-updated/master/setup-wifi-access-point.sh -O /home/${USER}/setup-wifi-access-point.sh
+else
+  touch /home/${USER}/setup-wifi-access-point.sh
+fi
 chmod 755 /home/${USER}/setup-wifi-access-point.sh
+
+# Install XLink Kai if XLINKKAI is true
+if [ "$XLINKKAI" = true ]; then
 
 # Remove old XLink Kai Repo if present
 sudo rm -rf /etc/apt/sources.list.d/teamxlink.list
@@ -81,23 +109,20 @@ while true; do
     sleep 5
 done
 EOF
+else
+touch /home/${USER}/launchkai.sh
 
+#End of XLink Kai install
+fi
 chmod 755 /home/${USER}/launchkai.sh
 
 # Install USB automount settings
-wget https://raw.githubusercontent.com/toolboc/psx-pi-smbshare/master/automount-usb.sh -O /home/${USER}/automount-usb.sh
+wget https://raw.githubusercontent.com/georgewoodall82/psx-pi-smbshare-updated/master/automount-usb.sh -O /home/${USER}/automount-usb.sh
 chmod 755 /home/${USER}/automount-usb.sh
-sudo /home/${USER}/automount-usb.sh
+/home/${USER}/automount-usb.sh
 
 # Set samba-init + ps3netsrv, wifi-to-eth-route, setup-wifi-access-point, and XLink Kai to run on startup
-{ echo -e "@reboot sudo bash /usr/local/bin/samba-init.sh\n@reboot sudo bash /home/${USER}/wifi-to-eth-route.sh && sudo bash /home/${USER}/setup-wifi-access-point.sh\n@reboot bash /home/${USER}/launchkai.sh"; } | crontab -u pi -
-
-# Start services
-sudo /usr/local/bin/samba-init.sh
-sudo /home/${USER}/wifi-to-eth-route.sh
-sudo /home/${USER}/setup-wifi-access-point.sh
-ps3netsrv++ -d /share/
-screen -dmS kailauncher /home/${USER}/launchkai.sh
+{ echo -e "@reboot sudo bash /usr/local/bin/samba-init.sh\n@reboot sudo bash /home/${USER}/wifi-to-eth-route.sh && sudo bash /home/${USER}/setup-wifi-access-point.sh\n@reboot bash /home/${USER}/launchkai.sh"; } | crontab -u ${USER} -
 
 # Not a bad idea to reboot
 sudo reboot
